Rearranged ragelGenerator header files
